split tests for activationresult
